prevent overwriting the nextcloud app directory
The  nextcloud application  directory includes  by default  the hosted
files and  the sqlite  database. The expectation  is that  these files
shall not be overwritten each  time the docker container is restarted.
The upstream configuration seems to have  a bug and does exactly this.
I disabled the installation and thus the potential removal of previous
app  data  completely. The  nextcloud  application  must be  installed
instead by the following commnad to be executed inside the container:

cp -R /usr/src/nextcloud /war/www/html/

Add the following configuration options to the config.php file:

    'datadirectory'     => '/var/www/html/nextcloud/data',
    'overwrite.cli.url' => '<hosturl>/nextcloud',

Eventually you will need to update the .htaccess file as well:

    ./occ maintenance:update:htaccess

Signed-off-by: Otto Linnemann <linneman@gmx.de>